refactor: 상품관리 메뉴 선택시 화면을 보여주는 함수 UI 로직으로 분리
