+++ src/js/components/manage/productAddMenu.js
@@ -1,8 +1,7 @@
-import { STORAGE_NAME } from '../../utils/constants.js';
 import { $ } from '../../utils/querySelector.js';
+import { STORAGE_NAME } from '../../utils/constants.js';
 import { isValidProductValue } from '../../utils/validation.js';
 import { setLocalStorage, getLocalStorage } from '../../utils/storage.js';
-import { productManageTemplate } from './manageTemplate.js';
 import { initProductManageScreen } from '../../view/view.js';
 
 const setProductItemsStorage = (productData) => {
@@ -12,7 +11,7 @@ const setProductItemsStorage = (productData) => {
   setLocalStorage(STORAGE_NAME.PRODUCT, storedProductItems);
 };
 
-const handleProductMenuSubmit = (event) => {
+export const handleProductMenuSubmit = (event) => {
   event.preventDefault();
   const productData = {
     name: $('#product-name-input').value.trim(),
@@ -26,15 +25,4 @@ const handleProductMenuSubmit = (event) => {
 
   initProductManageScreen(Array(productData));
   setProductItemsStorage(productData);
-};
-
-export const showProductAddMenu = () => {
-  $('#app-container').innerHTML = productManageTemplate;
-  const storedProductItems = getLocalStorage(STORAGE_NAME.PRODUCT);
-
-  if (storedProductItems) {
-    initProductManageScreen(storedProductItems);
-  }
-
-  $('form').addEventListener('submit', handleProductMenuSubmit);
 };
